+++ clients/admin-ui/src/app/store.ts
@@ -34,6 +34,7 @@ import {
 } from "user-management/index";
 
 import { STORAGE_ROOT_KEY } from "~/constants";
+import { reducer as featuresReducer } from "~/features/common/features";
 import { reducer as configWizardReducer } from "~/features/config-wizard/config-wizard.slice";
 import { scannerApi } from "~/features/config-wizard/scanner.slice";
 import {
@@ -104,6 +105,7 @@ const reducer = {
   dataUse: dataUseReducer,
   dataset: datasetReducer,
   datastoreConnections: datastoreConnectionReducer,
+  features: featuresReducer,
   organization: organizationReducer,
   subjectRequests: privacyRequestsReducer,
   system: systemReducer,